requirements.txt
Flask==3.0.2
replicate-python==0.22.2
openai==1.14.0
python-json-logger==2.0.7
requests==2.31.0